CMT 22/11 20.00

+++ app/src/main/java/com/example/smartwaiter/inteface/MenuItem.kt
@@ -1,3 +1,5 @@
 package com.example.smartwaiter.inteface
 
-data class MenuItem (var name:String,var description:String, var price:Float, var allergens:ArrayList<Int>, var image:String)
+data class MenuItem(
+    var name:String,
+    var description:String, var price: Double, var allergens:ArrayList<Int>, var image:String)
